fix: Move occupancy grid map filter (#748)
* create directory

* remove occupancy grid outlier filter from pointcloud_preprocessor

* rename package
