+++ sensing/pointcloud_preprocessor/CMakeLists.txt
@@ -49,7 +49,6 @@ ament_auto_add_library(pointcloud_preprocessor_filter SHARED
   src/outlier_filter/ring_outlier_filter_nodelet.cpp
   src/outlier_filter/voxel_grid_outlier_filter_nodelet.cpp
   src/outlier_filter/radius_search_2d_outlier_filter_nodelet.cpp
-  src/outlier_filter/occupancy_grid_map_outlier_filter_nodelet.cpp
   src/outlier_filter/dual_return_outlier_filter_nodelet.cpp
   src/passthrough_filter/passthrough_filter_nodelet.cpp
   src/passthrough_filter/passthrough_filter_uint16_nodelet.cpp
@@ -114,11 +113,6 @@ rclcpp_components_register_node(pointcloud_preprocessor_filter
 rclcpp_components_register_node(pointcloud_preprocessor_filter
   PLUGIN "pointcloud_preprocessor::RadiusSearch2DOutlierFilterComponent"
   EXECUTABLE radius_search_2d_outlier_filter_node)
-
-# -- Occupancy Grid Map Outlier Filter --
-rclcpp_components_register_node(pointcloud_preprocessor_filter
-  PLUGIN "pointcloud_preprocessor::OccupancyGridMapOutlierFilterComponent"
-  EXECUTABLE occupancy_grid_map_outlier_filter_node)
 
 # -- DualReturn Outlier Filter--
 rclcpp_components_register_node(pointcloud_preprocessor_filter
